third_party/blink/web_tests/fast/forms/fieldset/validation-in-fieldset-expected.txt

A form control in initially disabled fieldset:
PASS control1.willValidate is false
Then, enables the fieldset:
PASS document.getElementById("f1").disabled = false; control1.willValidate is true
A form control in initially enabled fieldset:
PASS control2.willValidate is true
Then, disables fieldset:
PASS fieldset2.disabled = true; control2.willValidate is false
Detach the form control from the fieldset:
PASS fieldset2.removeChild(control2); control2.willValidate is true
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
